update logging and add creating date update logic

+++ src/helpers/update-exif-metadata.ts
@@ -6,20 +6,36 @@ import { resolve } from 'path';
 
 const { unlink, copyFile } = fspromises;
 
-export async function updateExifMetadata(fileInfo: MediaFileInfo, timeTaken: string): Promise<void> {
+export async function updateExifMetadata(fileInfo: MediaFileInfo, timeTaken: string): Promise<void | 'notSupportError' | 'writeError' | 'updatedCreateDate'> {
   if (!doesFileSupportExif(fileInfo.mediaFilePath)) {
-    return;
+
+    console.log('---- notSupportError')
+
+    return 'notSupportError';
   }
 
   try {
+    if (fileInfo.mediaFileExtension === 'png') {
+      await exiftool.write(fileInfo.mediaFilePath, {
+        CreateDate: timeTaken,
+      });
+      return 'updatedCreateDate'
+    }
+
     await exiftool.write(fileInfo.mediaFilePath, {
       DateTimeOriginal: timeTaken,
     });
-  
+
     await unlink(`${fileInfo.mediaFilePath}_original`); // exiftool will rename the old file to {filename}_original, we can delete that
 
+    console.log('---- success')
+
   } catch (error) {
-    console.log('!!!!!!!!! error: ' + JSON.stringify(error) + ' \n' + fileInfo.mediaFilePath + ' ' + fileInfo.mediaFileName)
+
+    console.log('---- writeError')
+
+    return 'writeError'
+
     // await copyFile(fileInfo.mediaFilePath,  resolve(errorDir, fileInfo.mediaFileName));
     // if (fileInfo.jsonFileExists && fileInfo.jsonFileName && fileInfo.jsonFilePath) {
     //   await copyFile(fileInfo.jsonFilePath, resolve(errorDir, fileInfo.jsonFileName));
